+++ src/customizers/index.js
@@ -1,0 +1,7 @@
+const webpack = require("./webpack");
+const babel = require("./babel");
+
+module.exports = {
+  ...babel,
+  ...webpack
+};
